+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## [0.73.0] - 2025-03-07
+### Added
+- `reset-display-config`,
+- `restore-display-config`.
+### Changed
+- `reset-config` now resets display settings as well.
+
 ## [0.72.0] - 2025-03-07
 ### Added
 - `toggle-ephemeral`.